+++ src/api/ollama_client.rs
@@ -19,14 +19,17 @@ use serde_json::Value;
 use std::io;
 use std::process::Command;
 
+use crate::api::ChatApi;
+
 static LLM_PROTOCOL: &str = "http";
 static LLM_HOST: &str = "localhost";
 static LLM_PORT: &str = "11434";
 static LLM_ENDPOINT: &str = "/api/chat";
 
-pub(crate) struct OllamaClient {
+pub struct OllamaClient {
     model: String,
     pub(crate) system_prompt: String,
+    pub model_context_size: Option<usize>,
 }
 
 #[derive(Deserialize, Debug)]
@@ -41,27 +44,8 @@ pub(crate) struct OllamaMessage {
     pub(crate) content: String,
 }
 
-impl OllamaClient {
-    pub(crate) fn new(model: String, system_prompt: String) -> Self {
-        Self {
-            model,
-            system_prompt,
-        }
-    }
-
-    /// Send an empty message to ollama to preload the model.
-    pub(crate) fn verify(&self) -> io::Result<String> {
-        let send_body = serde_json::json!({
-            "model": self.model,
-        });
-
-        match Self::send_request_and_handle_response(&send_body) {
-            Ok(response) => Ok(response.message.content),
-            Err(e) => Err(e),
-        }
-    }
-
-    pub(crate) fn generate_response(
+impl ChatApi for OllamaClient {
+    fn generate_response(
         &self,
         history_messages_json: Value,
         user_prompt: &str,
@@ -77,6 +61,55 @@ impl OllamaClient {
 
         let response = Self::poll_for_response(&send_body)?;
         Ok(response.message.content)
+    }
+
+    fn model_context_size(&self) -> Option<usize> {
+        self.model_context_size
+    }
+
+    fn update_system_prompt(&mut self, new_system_prompt: String) {
+        self.system_prompt = new_system_prompt;
+    }
+}
+
+impl OllamaClient {
+    /// Create the client and verify that it is responding
+    pub fn new(model: String, system_prompt: String) -> Self {
+        Self {
+            model: model.clone(),
+            system_prompt,
+            model_context_size: None,
+        }
+    }
+
+    pub fn verify(&mut self) {
+        match self.preload() {
+            Ok(s) => println!("{s}"),
+            Err(e) => {
+                println!("\n\nModel is not available: {e}");
+                panic!(
+                    "Failed to verify ollama client\nCheck that Ollama is installed or run `ollama pull {}` to pull the model.",
+                    &self.model
+                );
+            }
+        }
+
+        self.model_context_size = Self::get_model_context_size(&self.model).unwrap_or_else(|e| {
+            eprintln!("Error getting model context size: {e}");
+            None
+        });
+    }
+
+    /// Send an empty message to ollama to preload the model.
+    fn preload(&self) -> io::Result<String> {
+        let send_body = serde_json::json!({
+            "model": self.model,
+        });
+
+        match Self::send_request_and_handle_response(&send_body) {
+            Ok(response) => Ok(response.message.content),
+            Err(e) => Err(e),
+        }
     }
 
     fn poll_for_response(send_body: &Value) -> io::Result<OllamaResponse> {
@@ -153,10 +186,6 @@ impl OllamaClient {
         messages.push(serde_json::json!({ "role": "user", "content": user_prompt }));
 
         messages
-    }
-
-    pub(crate) fn update_system_prompt(&mut self, new_system_prompt: String) {
-        self.system_prompt = new_system_prompt;
     }
 
     fn api_url() -> String {